chore: add duplex iterable type to connection
