+++ src/connection/utils.d.ts
@@ -1,0 +1,1 @@
+export function validateArgs(localAddr: any, localPeer: any, remotePeer: any, newStream: any, close: any, getStreams: any, stat: any): void;
